+++ theme/default/portfolio.php
@@ -1,13 +1,14 @@
 
-    <h1>Portfolio - beuster{se}</h1>
+    <h1>Portfolio</h1>
 <?php
-  if(count($data['portFiles']) > 0) {
-    foreach($data['portFiles'] as $portFileGroup) {
-      echo '<h2>'.$portFileGroup[0]['group'].'</h2>';
-      foreach($portFileGroup as $portFile) {
-        echo '  <div id="'.$portFile['id'].'">'."\n";
-        echo '    <img src="'.$portFile['path'].'" title="'.$portFile['name'].'" name="'.$portFile['name'].'" alt="'.$portFile['name'].'" class="portImage">'."\n";
-        echo '    <span style="display: none;">'.$portFile['text'].'</span>'."\n";
+  if(count($data['portSets']) > 0) {
+    foreach($data['portSets'] as $portSet) {
+      echo '<h2>'.$portSet->getName().'</h2>';
+      foreach($portSet->getItems() as $portItem) {
+        $portImage = $portItem->getImage();
+        echo '  <div>'."\n";
+        echo '    <img src="'.$portImage->getPath().'" title="'.$portImage->getTitle().'" name="'.$portImage->getTitle().'" alt="'.$portImage->getTitle().'" class="portImage">'."\n";
+        echo '    <span class="infotext">'.$portItem->getTitle().'</span>'."\n";
         echo '  </div>'."\n";
       }
     }